requirements.txt
streamlit>=1.38.0
serpapi>=1.0.0
agno>=0.3.6
fpdf2>=2.7.9
python-dotenv>=1.0.1
pandas>=2.2.2
requests>=2.32.3
